storage/requirements.txt
connexion==2.7.0
swagger-ui-bundle==0.0.8
SQLAlchemy==1.4.41
mysql-connector-python==8.0.30
pymysql==1.0.2
pykafka==2.4.0